renamed osm_stats package to galaxy #44
