+++ osm_galaxy.egg-info/SOURCES.txt
@@ -1,0 +1,13 @@
+README.md
+setup.py
+osm_galaxy.egg-info/PKG-INFO
+osm_galaxy.egg-info/SOURCES.txt
+osm_galaxy.egg-info/dependency_links.txt
+osm_galaxy.egg-info/requires.txt
+osm_galaxy.egg-info/top_level.txt
+src/galaxy/__init__.py
+src/galaxy/app.py
+src/galaxy/query_builder/__init__.py
+src/galaxy/query_builder/mapathon.py
+src/galaxy/validation/__init__.py
+src/galaxy/validation/mapathon.py
