+++ osm_galaxy.egg-info/requires.txt
@@ -1,0 +1,5 @@
+
+[dev]
+psycopg2
+pytest==3.7
+testing.postgresql==1.3.0
